refactor: typo
